Move the user avatar and DJ level command to more fitting categories
