+++ src/main/java/com/avairebot/commands/utility/UserAvatarCommand.java
@@ -19,15 +19,18 @@
  *
  */
 
-package com.avairebot.commands.administration;
+package com.avairebot.commands.utility;
 
 import com.avairebot.AvaIre;
 import com.avairebot.commands.CommandMessage;
 import com.avairebot.contracts.commands.Command;
+import com.avairebot.contracts.commands.CommandGroup;
+import com.avairebot.contracts.commands.CommandGroups;
 import com.avairebot.factories.MessageFactory;
 import com.avairebot.utilities.MentionableUtil;
 import net.dv8tion.jda.core.entities.User;
 
+import javax.annotation.Nonnull;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
@@ -64,6 +67,12 @@ public class UserAvatarCommand extends Command {
     @Override
     public List<String> getTriggers() {
         return Collections.singletonList("avatar");
+    }
+
+    @Nonnull
+    @Override
+    public List<CommandGroup> getGroups() {
+        return Collections.singletonList(CommandGroups.INFORMATION);
     }
 
     @Override
